requirements.txt
# FastAPI and server dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Auth0 and JWT dependencies
python-jose[cryptography]>=3.3.0
httpx>=0.25.0
cryptography>=41.0.0

# Optional: for async support
python-multipart>=0.0.6

# Database dependencies
sqlalchemy>=2.0.0
alembic>=1.13.0
psycopg2-binary>=2.9.9

# Supabase client library
supabase>=2.0.0

# Rate limiting
slowapi>=0.1.9

# Cryptography for encryption
cryptography>=41.0.0

# AWS S3 for audit logging
boto3>=1.34.0
botocore>=1.34.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
hypothesis>=6.92.0

# Kafka/Redpanda and Schema Registry
avro-python3>=1.11.0
confluent-kafka>=2.3.0
confluent-kafka[avro]>=2.3.0
